Metadata for page product, minor change on layout metadata
